backend/requirements.txt
absl-py @ file:///C:/ci/absl-py_1600290345082/work
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
argh==0.26.2
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
asn1crypto @ file:///C:/ci/asn1crypto_1594339244757/work
astor==0.8.1
astroid==2.3.3
astropy==4.0.1.post1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.9.1
bkcharts==0.2
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
blinker==1.4
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
clyent==1.2.2
colorama==0.4.3
comtypes==1.1.7
conda==4.8.5
conda-build==3.18.11
conda-package-handling==1.7.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cycler==0.10.0
cytoolz==0.10.1
dask @ file:///tmp/build/80754af9/dask-core_1600699676946/work
decorator==4.4.2
defusedxml==0.6.0
Django @ file:///tmp/build/80754af9/django_1600355803933/work
django-allauth==0.42.0
django-cors-headers==3.5.0
django-extensions==3.0.9
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
docutils==0.16
drf-yasg==1.17.1
entrypoints==0.3
Faker @ file:///tmp/build/80754af9/faker_1600355387751/work
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
flatbuffers==1.12
fsspec==0.7.4
future==0.18.2
gast==0.3.3
glob2==0.7
gmpy2==2.0.8
google-auth @ file:///tmp/build/80754af9/google-auth_1600960338579/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio @ file:///C:/ci/grpcio_1597406403308/work
h5py==2.10.0
HeapDict==1.0.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imagesize==1.2.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446525189/work
inflection==0.5.1
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1598883894321/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
jedi @ file:///C:/ci/jedi_1596472767194/work
Jinja2==2.11.2
joblib @ file:///tmp/build/80754af9/joblib_1594236160679/work
json5==0.9.5
jsonschema @ file:///C:/ci/jsonschema_1594363671836/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
jupyterlab==2.1.5
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
Keras==2.4.3
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.33.0+1.g022ab0f
locket==0.2.0
Markdown @ file:///C:/ci/markdown_1597415185889/work
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib @ file:///C:/ci/matplotlib-base_1597858387645/work
mccabe==0.6.1
menuinst==1.4.16
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.4.0
mpmath==1.1.0
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert @ file:///C:/ci/nbconvert_1594372737468/work
nbformat==5.0.7
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
nose @ file:///C:/ci/nose_1594377973236/work
notebook==6.0.3
numba==0.50.1
numexpr==2.7.1
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.4
pandas @ file:///C:/ci/pandas_1598371525254/work
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
partd==1.1.0
path==13.1.0
pathlib2 @ file:///C:/ci/pathlib2_1594381094851/work
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect @ file:///C:/ci/pexpect_1594383465186/work
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
Pillow @ file:///C:/ci/pillow_1594298234712/work
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
psutil==5.7.0
psycopg2 @ file:///C:/ci/psycopg2_1594305209841/work
py @ file:///tmp/build/80754af9/py_1593446248552/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyflakes==2.2.0
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
PyJWT==1.7.1
PyNaCl @ file:///C:/ci/pynacl_1595009241355/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyreadline==2.1
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
pytesseract==0.3.6
pytest==5.4.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes @ file:///C:/ci/pywin32-ctypes_1594392691209/work
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rope==0.17.0
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
Rtree==0.9.4
ruamel-yaml==0.15.87
ruamel.yaml.clib @ file:///C:/ci/ruamel.yaml.clib_1594419248566/work
scikit-image==0.16.2
scikit-learn @ file:///C:/ci/scikit-learn_1598376983131/work
scipy @ file:///C:/ci/scipy_1597686737426/work
seaborn==0.10.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
statsmodels==0.11.1
swagger-spec-validator==2.7.3
tb-nightly==2.4.0a20200907
tblib==1.6.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tf-estimator-nightly==2.4.0.dev2020090701
tf-nightly==2.4.0.dev20200907
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toolz==0.10.0
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1596810128862/work
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==1.35
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.9
xlwings==0.19.5
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
zipp==3.1.0
zope.event==4.4
zope.interface==4.7.1
